Allow people to identify themselves
Helps proving a point of awereness on keeping cookies using PHP's
default configuration which is using the server file system. As we have
two web machines serving, the cookie will only be present at one of the
machines making some of the requests (if made throw a load balancer
`www`) to not recover the cookie data.

The cookie being set defaults to the app domain, meaning that once
someone solves that problem all addresses available through this setup
will be able to recover the cookie peoperly.

Endpoints for identification accept HTML and plain text, so one can use
cURL (whith cookies) while presenting.